feat: creates init about us page
