+++ .vscode/settings.json
@@ -3,7 +3,7 @@
   "editor.formatOnSave": true,
   "editor.tabSize": 2,
   "editor.codeActionsOnSave": {
-    "source.fixAll": true
+    "source.fixAll": "explicit"
   },
   "headwind.runOnSave": false,
   "typescript.preferences.importModuleSpecifier": "non-relative"
