--- src/pages/projects/ProjectsPage.tsx
@@ -1,3 +1,0 @@
-const ProjectsPage = () => <div className=''>Projects Page</div>;
-
-export default ProjectsPage;
